functions-python/ratings/close_rating_round/requirements.txt
--extra-index-url https://_json_key_base64:ewogICJ0eXBlIjogInNlcnZpY2VfYWNjb3VudCIsCiAgInByb2plY3RfaWQiOiAibnV0bWVnLTkwOTljIiwKICAicHJpdmF0ZV9rZXlfaWQiOiAiYmY3M2M5ZDZiNjJhMDcxODcxMDM1MmZlZDVlNzgxOThlOTM2MDBiYiIsCiAgInByaXZhdGVfa2V5IjogIi0tLS0tQkVHSU4gUFJJVkFURSBLRVktLS0tLVxuTUlJRXZ3SUJBREFOQmdrcWhraUc5dzBCQVFFRkFBU0NCS2t3Z2dTbEFnRUFBb0lCQVFDMWhOMXBHSm1VN3JzQlxueVJ5cW94OFhRRzZBTlM0SzU2WklOdW9mNmcwbEdrWHhPNHo3MUs1eExHT055c2dLeXdCTU13aGQxQkpUUXhBTVxuTXJGSzhiWmZCT2M1VzROMWtOQnNQQndVTjFKemxjSzFMaUNyQ1R0V1BHMWJDT1loM0dQS1RuV1pMMmpCOExVcVxudU54Y0RVN21Pb1dtbXRua0g0TnRDUSthM1ZGYnJreDFudDNFS0lCV0xwandrSnV6a1podkdNMEc4MUhtU0svOVxuM05tZUFXRUtFSWhyRHdXQ2xJbUFsR3ZhS01JSFVENC9qTTZNU1FaOTB0eU1OSzVjYTZ4dXZFZ1J1ODNZTXR4RFxuK2wwUlo1ZUx6djRTREl0MmRON2ZVdUdPM3FqbnZLSlVaejR4NENaeTd0OTg0QmVWUnNXb0UrM3FWNkJUNHV1dlxuSHlaZ0ZvZXRBZ01CQUFFQ2dnRUFHRjY3OWl2MEZLdC9oblFZaWUxSUU1QldOWmJPTlc4WDQwcWlJTUpFM1FZYlxudlZFME1xeEJkbGJTRDl1WmcxOXhUUW10OWlEdWkraXRTeUgwTFI1cUVIUkJVZkZRSXpUYjBxZzRVZHVDa1p1WlxuN2tiQTRvMmsvb1V5M2wzQnZlcVBxRDNqNm5yVHV2UUU4clVxZXA1YnlTdXl2V1U5V3pRQzZsemswRE1tQk9PV1xuRG1Rd1dYQkNqN2V4OW1GM1hqNld2NC9NeEphcWNCVHQ3TEt2SEZTMTAvL0x6OFlVbTlsa2I1UndJVVl0eUVEWVxuSlBGN0EwV3hPMHlFb2s5S3ZUbHEzTXZsUTU0ZW1JZUxVK2oxNG1aRlVxTFErMXE2MDY2VGN4MTg4NTN3WjJldlxuaGVvcHJGM05idFhNSnpzaE9taGdXZWtwaHMzSWlKVGdyYUx4S2RDcjRRS0JnUURqdlAwb1dKaldqM3NORWk1OVxuemswQ2FLUUJwMWZEbUtsYzlkdlkwcVZTME9qVVErd3o3UkhXNTVqWkhTWWJYamtvYmxMb0ZSSiswZnVXRXYwTVxuQlJId21QQTlSZ3VmbzFCWHpza0NjNXI3eUdTVFpUTzNIUUJxMTU0UUN1UFJGRHFnclZrRk5BVHhVWVlnc1FIY1xuMXhDN0JkYy9qelZFQUMzY1lZcmxkdi9INFFLQmdRRE1DNG5EZzFRaWZmNGVkanQ4WWdPZjBZOWUxSlF6aXBLMFxuUlQwUmI3R1ByczMzTDVIR3F3eFE4TjR5RmZBSks4Z0xVOUtxUU44K3BuVldCNkhuQVVwRVpxc1ZQMmNPOWlGSlxuYnJ1S0lmZGJiZjBVMVNNdzdyTmpwcnYrNUxzNENKTVlPRm5OUEw3MGJ3VlFObW5GUGRrWEF6bVhiL2JyVVlqMVxuYkM4QVdUQ0pUUUtCZ1FDeldscFJsUHpCTlh4bVBlUnBaYmlXWmd5TDFOdTBpaDBQdmZ5bkZMWG9zK2c1aTBpUlxuV0dvTlFwWEluMEVzbXBjUE5WdUx0VlM4bGVpNk56Z2NhRDNKcXRqQW1OSjhwYVNYa3A0UmxUUUFpTUJVcUlZQ1xuOXc4bUZhdEEzU0t2dld1eXNQS0R1NFVNKy9SUytMUUwxVUF5MmhXNzA2L1NYd1cvSXY2OVdtdDBZUUtCZ1FDdFxuRHB3cFZwbGxKN2IvM3dkVTEweDdvQ0VaTFBNRlErSnBRdEFvbWJMZlNrTW1BeEtBalZKRTFBV2dOVHNBakVGK1xuV2ZteXJKS3JjRUxVVkdUY3FCdDFLQjZ5UnJRQjNIOGU2T3ZtbkNoUFExb0M4VUtIbmV4SCtJSTZEWVU4WVFsZ1xuNFdwSXg5eXRNL3d0cS9qSmNVdE9XVTNmUXZjSlhuM3FxWGF6dFBHVmdRS0JnUUNVMGpTeFZtTEx1T2U1Z084U1xueWwzeTF2bkU0VldaRGRBL3M2U09MekFBbmZTeGNvdDJtNmpBU0p6bzBNOTVwd3BUNUk3ZHdqNTltRFVUSWVMa1xuVlpmdStyWDZJL0Q2R2JteE1xVVZQaFVSL28zUHA0aTdMSUk5QnRmSnBYV2xmWFA5bTlRb2twNHo0T3JPQXJkSFxua09XM1BGWVdLVUZXL0oxZDdJV29FbkJCVUE9PVxuLS0tLS1FTkQgUFJJVkFURSBLRVktLS0tLVxuIiwKICAiY2xpZW50X2VtYWlsIjogIm51dG1lZy05MDk5Y0BhcHBzcG90LmdzZXJ2aWNlYWNjb3VudC5jb20iLAogICJjbGllbnRfaWQiOiAiMTA3NDk2OTg5NjczOTY1Nzk0NjQ5IiwKICAiYXV0aF91cmkiOiAiaHR0cHM6Ly9hY2NvdW50cy5nb29nbGUuY29tL28vb2F1dGgyL2F1dGgiLAogICJ0b2tlbl91cmkiOiAiaHR0cHM6Ly9vYXV0aDIuZ29vZ2xlYXBpcy5jb20vdG9rZW4iLAogICJhdXRoX3Byb3ZpZGVyX3g1MDlfY2VydF91cmwiOiAiaHR0cHM6Ly93d3cuZ29vZ2xlYXBpcy5jb20vb2F1dGgyL3YxL2NlcnRzIiwKICAiY2xpZW50X3g1MDlfY2VydF91cmwiOiAiaHR0cHM6Ly93d3cuZ29vZ2xlYXBpcy5jb20vcm9ib3QvdjEvbWV0YWRhdGEveDUwOS9udXRtZWctOTA5OWMlNDBhcHBzcG90LmdzZXJ2aWNlYWNjb3VudC5jb20iCn0K@us-central1-python.pkg.dev/nutmeg-9099c/quickstart-python-repo/simple/

firebase-admin==5.2.0
protobuf==3.20.1
py-firebase-dynamic-links==1.0.1
python-dateutil
nutmeg_utils==0.0.2

